Fixed generated database class
The database class static instance was not found through ressource.load due to a wrong path configuration in the generated DB class. You can now use the database class (i.e. the class that has a list to all instances of a class, for example if you have a character table, you will have a Character class with a scriptable object per row and a Character_DB class wit a scriptable objects that has a list with a reference to all the Character scriptable objects)